Multiple improvements to the rs_core_math CTS tests.
Range of possible values can now be specified for parameters.
To enable correct test generation for clamp, can tell that one
parameter should never be less than another.
Implemented many more test verifications.  Many are not yet
invoked because we need to change the way error is calculated.
That will be for the next CL.
Improved the random data generation for all types.

Change-Id: Iebb58611b35a2df633829e26d5634d4372c1dfc5